# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Trend musix
host: localhost:10010
basePath: /
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /chart/buildfeatures:
    x-swagger-router-controller: chart
    get:
      description: build chart features
      operationId: buildChartFeatures
      parameters:
        - name: startDate
          in: query
          description: start date MM/dd/yyyy
          type: string
        - name: endDate
          in: query
          description: end date MM/dd/yyyy
          type: string
        - name: genreType
          in: query
          description: genreType
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Response"  
  /chart/get-report:
    x-swagger-router-controller: chart
    get:
      description: get chart analysis report
      operationId: getChartReport
      parameters:
        - name: startDate
          in: query
          description: start date MM/dd/yyyy
          type: string
        - name: endDate
          in: query
          description: end date MM/dd/yyyy
          type: string
        - name: genreType
          in: query
          description: genreType
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Response"          
# complex objects have schema definitions
definitions:
...
  Response:
    properties:
      status:
        type: integer
      message:
        type: string